refactor: internal - changed Modal to LegacyModal (exported still as Modal) 7360977361
