+++ packages/core/src/components/LegacyModal/LegacyModal.module.scss
