+++ packages/core/src/components/LegacyModal/LegacyModalContent/LegacyModalContent.module.scss
