+++ packages/core/src/components/LegacyModal/LegacyModalFooter/LegacyModalFooter.module.scss
