+++ packages/core/src/components/LegacyModal/LegacyModalFooter/LegacyModalFooterButtons/__tests__/LegacyModalFooterButtons.snapshot.test.tsx
@@ -1,0 +1,26 @@
+import React from "react";
+import renderer from "react-test-renderer";
+import ModalFooterButtons from "../LegacyModalFooterButtons";
+
+describe("ModalFooterButtons", () => {
+  it("should render correctly", () => {
+    const tree = renderer
+      .create(<ModalFooterButtons primaryButtonText="Confirm" secondaryButtonText="Cancel" />)
+      .toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+
+  it("should render correctly with no tertiary", () => {
+    const tree = renderer.create(<ModalFooterButtons primaryButtonText="Confirm" />).toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+
+  it("should render correctly with disabled primary", () => {
+    const tree = renderer
+      .create(
+        <ModalFooterButtons primaryButtonText="Confirm" secondaryButtonText="Cancel" disablePrimaryButton={true} />
+      )
+      .toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});
