+++ packages/core/src/components/LegacyModal/LegacyModalHeader/LegacyModalHeader.module.scss
