+++ packages/core/src/components/LegacyModal/__stories__/LegacyModal.stories.module.scss
