+++ packages/core/src/components/LegacyModal/a11yDialog.ts
