+++ packages/core/src/components/LegacyModal/index.ts
@@ -1,0 +1,19 @@
+export { default as Modal, LegacyModalProps as ModalProps } from "./LegacyModal";
+export {
+  default as ModalHeader,
+  LegacyModalHeaderProps as ModalHeaderProps
+} from "./LegacyModalHeader/LegacyModalHeader";
+export {
+  default as ModalContent,
+  LegacyModalContentProps as ModalContentProps
+} from "./LegacyModalContent/LegacyModalContent";
+export {
+  default as ModalFooter,
+  LegacyModalFooterProps as ModalFooterProps
+} from "./LegacyModalFooter/LegacyModalFooter";
+export {
+  default as ModalFooterButtons,
+  LegacyModalFooterButtonsProps as ModalFooterButtonsProps
+} from "./LegacyModalFooter/LegacyModalFooterButtons/LegacyModalFooterButtons";
+
+export * from "./LegacyModal.types";
