+++ packages/core/src/components/LegacyModal/useAnimationProps.ts
