--- packages/core/src/components/Modal/ModalContent/__tests__/ModalContent.snapshot.test.tsx
@@ -1,21 +1,0 @@
-import React from "react";
-import renderer from "react-test-renderer";
-import ModalContent from "../ModalContent";
-
-describe("ModalContent renders correctly", () => {
-  it("with empty props", () => {
-    const tree = renderer.create(<ModalContent>content</ModalContent>).toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-
-  it("with children", () => {
-    const tree = renderer
-      .create(
-        <ModalContent>
-          <div>children</div>
-        </ModalContent>
-      )
-      .toJSON();
-    expect(tree).toMatchSnapshot();
-  });
-});
