--- packages/core/src/components/Modal/index.ts
@@ -1,10 +1,0 @@
-export { default as Modal, ModalProps } from "./Modal";
-export { default as ModalHeader, ModalHeaderProps } from "./ModalHeader/ModalHeader";
-export { default as ModalContent, ModalContentProps } from "./ModalContent/ModalContent";
-export { default as ModalFooter, ModalFooterProps } from "./ModalFooter/ModalFooter";
-export {
-  default as ModalFooterButtons,
-  ModalFooterButtonsProps
-} from "./ModalFooter/ModalFooterButtons/ModalFooterButtons";
-
-export * from "./Modal.types";
